// packages/focalm-server/src/generated/db-client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/**
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/
export const TaskState = {
  Pending: 'Pending',
  Active: 'Active',
  Done: 'Done'
} as const

export type TaskState = (typeof TaskState)[keyof typeof TaskState]


export const TaskHistoryOperation = {
  Set: 'Set',
  Add: 'Add',
  Remove: 'Remove'
} as const

export type TaskHistoryOperation = (typeof TaskHistoryOperation)[keyof typeof TaskHistoryOperation]


export const TaskHistoryKey = {
  title: 'title',
  state: 'state',
  archived: 'archived',
  impact: 'impact',
  ease: 'ease',
  authorId: 'authorId',
  responsibleId: 'responsibleId',
  participants: 'participants',
  relations: 'relations',
  orderKey: 'orderKey',
  parentId: 'parentId',
  projectId: 'projectId',
  description: 'description',
  startAfterDate: 'startAfterDate',
  startAfterOffset: 'startAfterOffset',
  plannedStartDate: 'plannedStartDate',
  plannedStartOffset: 'plannedStartOffset',
  dueToDate: 'dueToDate',
  dueToOffset: 'dueToOffset'
} as const

export type TaskHistoryKey = (typeof TaskHistoryKey)[keyof typeof TaskHistoryKey]


export const CreatedAtFixReason = {
  Low: 'Low',
  High: 'High',
  Both: 'Both'
} as const

export type CreatedAtFixReason = (typeof CreatedAtFixReason)[keyof typeof CreatedAtFixReason]


export const PermissionKind = {
  read: 'read',
  update: 'update',
  create: 'create',
  delete: 'delete'
} as const

export type PermissionKind = (typeof PermissionKind)[keyof typeof PermissionKind]


export const PermissionInProject = {
  tasks: 'tasks',
  participants: 'participants'
} as const

export type PermissionInProject = (typeof PermissionInProject)[keyof typeof PermissionInProject]